Make browser instantiation more resilient.
The underlying selenium function that results in the "Can't load profile"
error is a 30-second polling function. Since that is also our default
timeout setting, bok-choy never gets a chance to retry. (That is, it takes
30 seconds for selenium to produce the WebDriverException). This gives
the driver more chances to create a browser instance when previously it
would only try once and fail. That was not the intention of the original
change with a `try_limit`. This also overrides the default Promise timeout.