Update Sphinx docs to edX theme
PLAT-1087 - Update the Sphinx docs to use the theme from the  edx-sphinx-theme repo. Also ran into an issue with the docs tests that is unrelated but may have failed on commit so I fixed that too. Problem just seems to have been that GitHub changed their html layout. Also ran into issues with needle has a max version of selenium so I had to peg the version in setup.py to prevent a version error when applying requirements when installing through pip.