Update Sphinx docs to edX theme, fix doc test (#179)
* Update Sphinx docs to edX theme, fix doc test

PLAT-1087 - Update the Sphinx docs to use the theme from the  edx-sphinx-theme repo. Also ran into an issue with the docs tests that is unrelated but may have failed on commit so I fixed that too. Problem just seems to have been that GitHub changed their html layout. Also ran into issues where needle has a max version of selenium so I had to limit the version in setup.py to prevent a version error when installing that way.